requirements.txt
streamlit
langchain_groq
langchain_community
pandas
python-dotenv
tabulate
